moved ClickOnLayerActivity to query section
